qwen_fastrtc/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
fastrtc[vad,stt,tts]==0.0.32
transformers>=4.52.3
torch
numpy
python-dotenv
